Create barcode_fasta_generator.py
PacBio CLI demultiplexing tools require a fasta file that lists the barcodes used in a sequencing project.  While it is possible to run the tools with the entire list of barcodes, this generates a lot of empty files for the barcodes that were not used in the project.

I created this tool to output a fasta file containing only the barcode names that were selected by the user.